Struct tcl::object::Object

pub struct Object<'env> {
    // some fields omitted
}

A Tcl value

Methods

impl<'env> Object<'env>

fn from_raw(env: &TclEnvironment, raw: *mut Tcl_Obj) -> Object

fn new(env: &TclEnvironment) -> Object

Create a new untyped Tcl value

fn new_boolean(env: &TclEnvironment, val: bool) -> Object

Create a new boolean Tcl value

fn new_integer(env: &TclEnvironment, val: i32) -> Object

Create a new integer Tcl value

fn new_long(env: &TclEnvironment, val: i64) -> Object

Create a new long Tcl value

fn new_double(env: &TclEnvironment, val: f64) -> Object

Create a new double Tcl value

fn new_string(env: &'env TclEnvironment, val: &str) -> Object<'env>

Create a new string Tcl value

fn new_byte_array(env: &'env TclEnvironment, val: &[u8]) -> Object<'env>

Create a new byte array Tcl value

fn set_boolean(&mut self, val: bool)

Set the contents of a Tcl value to a boolean

fn set_integer(&mut self, val: i32)

Set the contents of a Tcl value to an integer

fn set_long(&mut self, val: i64)

Set the contents of a Tcl value to a long

fn set_double(&mut self, val: f64)

Set the contents of a Tcl value to a double

fn set_string(&mut self, val: &str)

Set the contents of a Tcl value to a string

fn set_byte_array(&mut self, val: &[u8])

Set the contents of a Tcl value to a byte array

unsafe fn raw(&self) -> *mut Tcl_Obj

fn get_string(&self) -> String

Get the string representation of a Tcl value

fn get_byte_array(&self) -> Vec<u8>

Get the byte array representation of a Tcl value

fn is_shared(&self) -> bool

Is the value currently used to represent multiple variables in an interpreter