Enum tcl::tcl::TclResult

pub enum TclResult {
    Ok,
    Error(String),
    Return,
    Break,
    Continue,
}

Represents the possible states an interpreter can be in after an evaluation

Variants

Ok

The command completed successfully

Error

There was an error

Return

The last command was a Return from a function

Break

The last command was a Break from a loop

Continue

The last command was a Continue from a loop

Methods

impl TclResult

fn from_ll(result: i32, interpreter: &Interpreter) -> TclResult

Convert a low-level representation of a Tcl result into a TclResult enum instance

fn is_ok(&self) -> bool