Struct tcl::object::Object

pub struct Object<'env> {
    // some fields omitted
}

A Tcl value

Methods

impl<'env> Object<'env>

fn from_raw(env: &TclEnvironment, raw: *mut Tcl_Obj) -> Object

fn new<V: TclObject>(env: &TclEnvironment, val: V) -> Object

Create a new Tcl value

fn set<V: TclObject>(&mut self, val: V)

Set the contents of this Tcl object to val

fn get<'a, V: TclObject>(&self, interp: &'a mut Interpreter) -> Result<V, Cow<'a, str>>

Get the contents of this Tcl object

fn get_string(&self) -> String

Get the string representation of a Tcl value

fn get_byte_array(&self) -> Vec<u8>

Get the byte array representation of a Tcl value

unsafe fn raw(&self) -> *mut Tcl_Obj

fn is_shared(&self) -> bool

Is the value currently used to represent multiple variables in an interpreter